app/worker/requirements.txt
azure-servicebus==7.11.4
azure-storage-blob==12.19.0
azure-data-tables==12.4.4
prometheus-client==0.19.0
pytesseract==0.3.10
Pillow==10.1.0